Improvements: time now is a u32 for better padding

+++ prototypes/src/prototype/event.rs
@@ -66,4 +66,4 @@ impl<I> Event<I> {
     }
 }
 
-pub type Time = u64;
+pub type Time = u32;
